# [c10d] Faster coalescing