--- a/torch/distributed/__init__.py
+++ b/torch/distributed/__init__.py
@@ -68,6 +68,8 @@ if is_available():
         _create_process_group_wrapper,
         _rank_not_in_group,
         _c10d_error_logger,
+        _coalescing_manager,
+        _CoalescingManager,
     )
 
     from .rendezvous import (