@kwen2501 2023-04-11
Thanks for the comments @kumpera.
> This design is problematic as it silently breaks the contract of collectives which is to return a Work object if `async_op=True`.

I think there is no precedented contract when it comes to a context manager.

> It means we lose the capability to use `_coalescing_manager ` to wrap a chunk of code to transparently coalesce their collectives.

Calling `work.wait()` inside the context manager is ineffective and can be deemed illegal.
So transparently moving a chunk of existing code inside the context manager is not supported.
Returning `None` from `dist.all_reduce` prevent users from making these false calls because it would error out.