@kwen2501 2023-04-12
> The contract we must follow is the one documented for collectives. Having context managers changing basic assumptions of how a function operate is a bad design.

I can return a Work object that is illegal to wait on.
In any case, I don't think updating the document to add "riders" for use under new circumstances is a bad idea.

> to wrap a chunk of code to transparently coalesce their collectives

Btw, this is exactly one advantage of context manager over asking users to construct a list of op representations.
I am understanding "a chunk of code" as "a chunk of (existing) code".