@kumpera 2023-04-13
> Thanks for the feedback @kumpera.
> 
> > It needs to return either one Work instance of have len(reqs) be equal to the number of collectives - even if there are duplicate Work objects.
> 
> IIUC, "of" -> "or".
> 
> I am thinking of going with @mrshenli 's suggestion above:
> 
> ```
> with _coalescing_manager() as cm:
>   ...
> 
> cm.wait()
> ```
> 
> Would that satisfy the requirement before "or"?
> 
> I am avoiding returning a number of requests corresponding to the number of collectives because:
> 
> ```
> for req in reqs:
>     req.wait()
> ```
> 
> would also induce a lot of binding overhead. And in current case, they are duplicated calls. Also, since the user has asked to coalesce the works, it should be deemed a single work.

That would work great. One work handle is better than nothing.

BTW, if you rather have it return torch.futures.Future instead, I'm all for it as part of our slow paced effort of deprecating Work.